feat: relocate draft paths
