+++ src/apis/online/drafts/index.ts
@@ -1,4 +1,4 @@
-import { AuthToken } from "../../../@types";
+import { AuthToken } from "../../@types";
 import { Draft, DraftParameter, NewDraft } from "./@types";
 
 export interface Methods {
