+++ src/index.ts
@@ -12,7 +12,7 @@ export * from "./apis/blog/revisions/@types";
 export * from "./apis/blog/categories/@types";
 export * from "./apis/blog/revisions/contrib/@types";
 export * from "./apis/online/exhibition/@types";
-export * from "./apis/online/exhibition/drafts/@types";
+export * from "./apis/online/drafts/@types";
 export * from "./apis/onsite/exhibition/status/@types";
 export * from "./apis/onsite/general/term/@types";
 export * from "./apis/onsite/general/guest/@types";
